Rename DotNotationWrapper() -> ItemSpec()
And move from utils.py to items.py.

No changes in behavior were made. That comes later.
