+++ onyo/cli/tests/test_get.py
@@ -13,7 +13,10 @@ from onyo.lib.consts import (
 )
 from onyo.lib.onyo import OnyoRepo
 from onyo.lib.command_utils import natural_sort
-from onyo.lib.utils import DotNotationWrapper
+from onyo.lib.items import (
+    Item,
+    ItemSpec,
+)
 
 if TYPE_CHECKING:
     from typing import (
@@ -27,7 +30,6 @@ def convert_contents(
         raw_assets: list[tuple[str, dict[str, Any]]]) -> Generator:
     r"""Convert content dictionary to a plain-text string."""
 
-    from onyo.lib.items import Item
     for file, raw_contents in raw_assets:
         yield [file, Item(raw_contents).yaml()]
 
@@ -590,7 +592,7 @@ def test_get_keys(repo: OnyoRepo,
 
     # Get all the key values and make sure they match
     for line in output:
-        asset = DotNotationWrapper(raw_assets[[a[0] for a in raw_assets].index(line[-1])][1])
+        asset = ItemSpec(raw_assets[[a[0] for a in raw_assets].index(line[-1])][1])
 
         for i, key in enumerate(keys):
             if key in PSEUDO_KEYS:
@@ -771,13 +773,13 @@ def test_natural_sort(keys: dict[str, sort_t],
                       expected: list[int]) -> None:
     r"""Test implementation of natural sorting algorithm."""
 
-    assets = [DotNotationWrapper(t[1]) for t in asset_contents
+    assets = [ItemSpec(t[1]) for t in asset_contents
               if t[0] in ['a13bc_foo_bar.1',
                           'a2cd_foo_bar.2',
                           'a36ab_foo_bar.3',
                           'a36ab_afoo_bar.4']]
     sorted_assets = natural_sort(assets, keys=keys)  # pyre-ignore[6]
-    # ^ No idea why this is the only place where pyre can't figure that DotNotationWrapper is a UserDict
+    # ^ No idea why this is the only place where pyre can't figure that ItemSpec is a UserDict
     assert expected == [data.get('id') for data in sorted_assets]
 
     # explicitly check path sorting:
