+++ onyo/lib/items.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from collections import UserDict
 from copy import deepcopy
 from pathlib import Path
 from typing import TYPE_CHECKING
@@ -19,7 +20,6 @@ from onyo.lib.pseudokeys import (
     PseudoKey,
 )
 from onyo.lib.utils import (
-    DotNotationWrapper,
     dict_to_yaml,
 )
 
@@ -27,6 +27,7 @@ from onyo.lib.utils import (
 if TYPE_CHECKING:
     from typing import (
         Any,
+        Generator,
         Mapping,
         TypeVar,
     )
@@ -44,7 +45,179 @@ def resolve_alias(key: Any) -> Any:
         return key
 
 
-class Item(DotNotationWrapper):
+class ItemSpec(UserDict):
+    """Access nested dictionaries via hierarchical keys.
+
+    Wrap a dictionary to traverse multidimensional dictionaries using a dot as
+    the delimiter. In other words, it provides a view of the flattened
+    dictionary::
+
+      > d = {'key': 'value', 'nested': {'key': 'another value'}}
+      > wrapper = ItemSpec(d)
+      > wrapper['nested.key']
+      'another value'
+      > list(wrapper.keys())
+      ['key', 'nested.key']
+
+    Iteration only considers the flattened view. Keys that contain a dictionary
+    are not yielded when using ``.keys()``, ``.values()``, and ``.items()``.
+
+    The underlying dictionary is available via the ``.data`` attribute when the
+    standard Python behavior is needed.
+    """
+
+    def __init__(self,
+                 __dict: Mapping[_KT, _VT] | None = None,
+                 pristine_original: bool = True,
+                 **kwargs: _VT) -> None:
+        r"""Initialize a dot notation wrapped dictionary.
+
+        Parameters
+        ----------
+        __dict
+            Dictionary to wrap.
+        pristine_original
+            Store ``__dict`` unaltered in the ``.data`` attribute.
+            This behavior is the primary intended use for the wrapper: just a
+            namespace wrapper for dicts.
+
+            Set to ``False`` to interpret the incoming dict's keys for dot
+            notation and set accordingly.
+        """
+
+        if pristine_original and __dict and isinstance(__dict, dict):
+            # Maintain the original dict object (and class).
+            # NOTE: Would modify wrapped dict w/ kwargs if both are given.
+            #       deepcopy would prevent this, but contradicts the idea of wrapping.
+            super().__init__()
+            self.data = __dict
+            self.update(**kwargs)
+        else:
+            # Resort to `UserDict` behavior.
+            super().__init__(__dict, **kwargs)
+
+    def _keys(self) -> Generator[str, None, None]:
+        """Yield all keys recursively from nested dictionaries in dot notation.
+
+        A by-product of dot notation is that all keys are strings, regardless of
+        their original type in the underlying dictionary.
+
+        Keys that contain a dictionary not yielded.
+        """
+
+        def recursive_keys(d: dict):
+            for k in d.keys():
+                if hasattr(d[k], "keys"):
+                    yield from (k + "." + sk for sk in recursive_keys(d[k]))
+                else:
+                    # Cast as a string. One can't have a key 'some.1.more',
+                    # where 1 remains an integer.
+                    yield str(k)
+
+        yield from recursive_keys(self.data)
+
+    def __getitem__(self,
+                    key: _KT) -> Any:
+        r"""Get the value of a key."""
+
+        if isinstance(key, str):
+            parts = key.split('.')
+            effective_dict = self.data
+            if len(parts) > 1:
+                for lvl in range(len(parts) - 1):
+                    try:
+                        effective_dict = effective_dict[parts[lvl]]
+                    except KeyError as e:
+                        raise KeyError(f"'{'.'.join(parts[:lvl + 1])}'") from e
+                    except TypeError as e:
+                        raise KeyError(f"'{'.'.join(parts[:lvl])}' is not a dictionary.") from e
+
+            try:
+                return effective_dict[parts[-1]]
+            except KeyError as e:
+                raise KeyError(f"'{key}'") from e
+            except TypeError as e:
+                raise KeyError(f"'{'.'.join(parts[:-1])}' is not a dictionary.") from e
+
+        return super().__getitem__(key)
+
+    def __setitem__(self,
+                    key: _KT,
+                    item: _VT) -> None:
+        r"""Set a key.
+
+        Keys that are strings are interpreted for dot notation, and intermediate
+        dictionaries are created as needed.
+        """
+
+        if isinstance(key, str):
+            parts = key.split('.')
+            effective_dict = self.data
+            if len(parts) > 1:
+                for lvl in range(len(parts) - 1):
+                    try:
+                        effective_dict = effective_dict[parts[lvl]]
+                    except KeyError:
+                        # nested dict doesn't exist yet
+                        effective_dict[parts[lvl]] = dict()
+                        effective_dict = effective_dict[parts[lvl]]
+
+            effective_dict[parts[-1]] = item
+        else:
+            super().__setitem__(key, item)
+
+    def __delitem__(self,
+                    key: _KT) -> None:
+        r"""Remove a ``key`` from self."""
+
+        if isinstance(key, str):
+            parts = key.split('.')
+            effective_dict = self.data
+            if len(parts) > 1:
+                for lvl in range(len(parts) - 1):
+                    try:
+                        effective_dict = effective_dict[parts[lvl]]
+                    except KeyError as e:
+                        raise KeyError(f"'{'.'.join(parts[:lvl + 1])}'") from e
+            del effective_dict[parts[-1]]
+        else:
+            super().__delitem__(key)
+
+    def __contains__(self,
+                     key: _KT) -> bool:
+        """Whether ``key`` is in self.
+
+        Unlike iteration over keys, keys that contain a dictionary are
+        matchable and return ``True``.
+        """
+
+        try:
+            self.__getitem__(key)
+            return True
+        except KeyError:
+            return False
+
+    def __iter__(self) -> Generator[str, None, None]:
+        r"""Return the iterator.
+
+        A by-product of dot notation is that all keys are strings, regardless of
+        their original type in the underlying dictionary.
+
+        Keys that contain a dictionary are not yielded.
+        """
+
+        return self._keys()
+
+    def __len__(self) -> int:
+        r"""Return the number of keys in the dot notation view.
+
+        Keys that contain a dictionary are not counted.
+        """
+
+        return len(list(self._keys()))
+
+
+class Item(ItemSpec):
     r"""An item that an :py:class:`onyo.lib.inventory.Inventory` can potentially track.
 
     The main purpose is to attach pseudo-keys and alias resolution to things
