+++ onyo/lib/tests/test_items_ItemSpec.py
@@ -1,5 +1,5 @@
 import pytest
-from onyo.lib.utils import DotNotationWrapper
+from onyo.lib.items import ItemSpec
 
 
 def test_init_pristine_original():
@@ -9,7 +9,7 @@ def test_init_pristine_original():
          'nested.deep.key': 'value',
          }
 
-    wrapper = DotNotationWrapper(d, pristine_original=False)
+    wrapper = ItemSpec(d, pristine_original=False)
     assert wrapper.data['some'] == d['some']
     assert 'nested' in wrapper.data and isinstance(wrapper.data['nested'], dict)
     assert wrapper.data['nested']['one'] == d['nested.one']
@@ -27,7 +27,7 @@ def test_get_values():
                     }
          }
 
-    wrapper = DotNotationWrapper(d)
+    wrapper = ItemSpec(d)
     assert wrapper['some'] == d['some']
     assert wrapper['nested'] == d['nested']  # TODO: check dict-equal-helper
     assert wrapper['nested.one'] == d['nested']['one']
@@ -62,7 +62,7 @@ def test_set_values():
                     }
          }
 
-    wrapper = DotNotationWrapper(d)
+    wrapper = ItemSpec(d)
     wrapper['some'] = 'newvalue'
     assert wrapper.get('some') == 'newvalue'
     assert d['some'] == 'newvalue'
@@ -83,7 +83,7 @@ def test_set_values():
 
     # update from another wrapped dict should allow for "recursive update":
     updater = {'nested': {'one': 3}}
-    wrapper.update(DotNotationWrapper(updater))
+    wrapper.update(ItemSpec(updater))
     assert wrapper['nested.one'] == updater['nested']['one']
     assert wrapper['nested.two'] == '2'
 
@@ -98,7 +98,7 @@ def test_magic_methods():
          }
     dot_keys = ['some', 'nested.one', 'nested.two', 'nested.deep.key']
 
-    wrapper = DotNotationWrapper(d)
+    wrapper = ItemSpec(d)
     # .keys()
     keys = [k for k in wrapper.keys()]
     assert all(k in keys for k in dot_keys)
